Move `Node` type into `node` module
